Refactor Token generation and verification
