+++ app/Http/Controllers/Auth/AuthenticatedSessionController.php
@@ -5,24 +5,25 @@ namespace App\Http\Controllers\Auth;
 use App\Models\v1\JwtToken;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
-use App\Services\TokenService;
 use Illuminate\Http\JsonResponse;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\Auth\LoginRequest;
+use App\Services\Jwt\Interfaces\TokenGenerator;
+
 
 class AuthenticatedSessionController extends Controller
 {
     /**
      * Handle an incoming authentication request.
      */
-    public function store(LoginRequest $request, TokenService $tokenService): JsonResponse
+    public function store(LoginRequest $request, TokenGenerator $tokenGenerator): JsonResponse
     {
         $request->authenticate();
 
         $user = $request->user();
         $uniqueId = bin2hex(random_bytes(8));
-        $token = $tokenService->generateToken($user->uuid, $uniqueId);
+        $token = $tokenGenerator->generateToken($user->uuid, $uniqueId);
 
         try {
             $this->createOrUpdateJwtToken($user, $uniqueId);
