+++ app/Services/Jwt/Interfaces/TokenGenerator.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Services\Jwt\Interfaces;
+
+interface TokenGenerator
+{
+    public function generateToken(string $userUuid, string $uniqueId): string;
+}
