+++ app/Services/Jwt/Interfaces/TokenVerifier.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Services\Jwt\Interfaces;
+
+interface TokenVerifier
+{
+    public function verifyToken(string $jwt): ?string;
+}
