+++ app/Services/Jwt/SystemClock.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Services\Jwt;
+
+use Psr\Clock\ClockInterface;
+
+class SystemClock implements ClockInterface
+{
+    public function now(): \DateTimeImmutable
+    {
+        return new \DateTimeImmutable();
+    }
+}
