+++ composer.json
@@ -15,6 +15,7 @@
         "laravel/sanctum": "^3.2",
         "laravel/tinker": "^2.8",
         "lcobucci/jwt": "^5.0",
+        "psr/clock": "^1.0",
         "spatie/laravel-sluggable": "^3.4"
     },
     "require-dev": {
